Extract common hub logic out
