+++ Octoprotecto/GameHub.cs
@@ -5,7 +5,7 @@ using Microsoft.Extensions.Logging;
 
 namespace Octoprotecto
 {
-    public class GameHub : Common.GameHub<OctoprotectoRoom, Octopus>
+    public partial class GameHub : Common.GameHub<OctoprotectoRoom, Octopus>
     {
         public GameHub(OctoprotectoLobby lobby) 
             : base(lobby)
@@ -34,14 +34,14 @@ namespace Octoprotecto
             Logger.LogInformation("PLAYER JOINS ROOM:{0}.", roomId);
             if (!this.GameLobby.Rooms.ContainsKey(roomId))
             {
-                await Clients.Caller.SendAsync("ErrorJoiningRoom", $"Room {roomId} not found.");
+                await this.JoinRoomError($"Room {roomId} not found.");
                 return;
             }
 
             var room = this.GameLobby.Rooms[roomId];
             if (room.State != OctoprotectoRoom.RoomState.SettingUp)
             {
-                await Clients.Caller.SendAsync("ErrorJoiningRoom", $"Room {roomId} already started.");
+                await this.JoinRoomError($"Room {roomId} already started.");
                 return;
             }
 
@@ -49,7 +49,7 @@ namespace Octoprotecto
             var playerId = Utils.GenerateId(10, allKeys);
             if (playerId == null)
             {
-                await Clients.Caller.SendAsync("ErrorJoiningRoom", "Error generating player ID.");
+                await this.JoinRoomError("Error generating player ID.");
                 return;
             }
 
@@ -62,16 +62,7 @@ namespace Octoprotecto
             if (soloRun
                 && string.IsNullOrEmpty(roomId))
             {
-                var newRoom = this.GameLobby.CreateRoom(Context.ConnectionId);
-                if (newRoom == null)
-                {
-                    await Clients.Caller.SendAsync(this.Message_ShowError, $"Room {roomId} cannot be started.");
-                    return;
-                }
-
-                await Clients.Caller.SendAsync("RoomCreated", newRoom.RoomId);
-                newRoom.CreatePlayer("SoloPlayer", Context.ConnectionId);
-                newRoom.StartGame();
+                await this.StartSoloRun(roomId);
                 return;
             }
 
@@ -115,29 +106,12 @@ namespace Octoprotecto
                 }
 
                 // Re-connect a player
-                await Groups.AddToGroupAsync(Context.ConnectionId, roomId);
-                await Groups.RemoveFromGroupAsync(octopus.ConnectionId, roomId);
-                octopus.ConnectionId = Context.ConnectionId;
-                await Clients.Caller.SendAsync("InitializeNewPlayer", room.RoomId, room.OctopiMovementBounds, octopus);
+                await this.ReconnectPlayer(room, octopus);
                 return;
             }
 
             // Re-connect the host room
-            if (room.State != OctoprotectoRoom.RoomState.SettingUp)
-            {
-                await Clients.Caller.SendAsync(this.Message_ShowError, $"Room {roomId} has already started. Resuming is not supported.", true /*Refresh on click*/);
-                return;
-            }
-
-            room.ConnectionId = Context.ConnectionId;
-            Logger.LogInformation("Reconnected to Room {0}", room.RoomId);
-            await Clients.Caller.SendAsync("RoomCreated", room.RoomId);
-            foreach (var roomPlayer in room.Players)
-            {
-                var roomOctopus = roomPlayer.Value;
-                var roomPlayerId = roomPlayer.Key;
-                await Clients.Caller.SendAsync("SpawnOctopus", roomOctopus);
-            }
+            await this.ReconnectHost(room);
         }
 
         public async Task UpdateOctopusPosition(string roomId, string playerId, double x, double y)
@@ -182,31 +156,6 @@ namespace Octoprotecto
 
             await Clients.Client(room.ConnectionId).SendAsync("SpawnOctopus", octopus);
             await Clients.Client(octopus.ConnectionId).SendAsync("OctopusRespawn", room.OctopiMovementBounds, octopus);
-        }
-
-        private async Task CreateNewOctopus(OctoprotectoRoom room, string playerId, string colorIn)
-        {
-            if (playerId == null)
-            {
-                await Clients.Caller.SendAsync(this.Message_ShowError, $"Invalid ID provided.");
-                return;
-            }
-
-            var octopus = room.CreatePlayer(playerId, Context.ConnectionId);
-
-            if (octopus == null)
-            {
-                await Clients.Caller.SendAsync(this.Message_ShowError, $"Unable to create player.");
-                return;
-            }
-
-            // Set additional fields
-            var color = int.Parse(colorIn.TrimStart('#'), System.Globalization.NumberStyles.HexNumber);
-            octopus.Tint = color;
-            octopus.SetRandomLocation(room.OctopiMovementBounds);
-
-            await Clients.Caller.SendAsync("InitializeNewPlayer", room.RoomId, room.OctopiMovementBounds, octopus);
-            await Clients.Client(room.ConnectionId).SendAsync("SpawnOctopus", octopus);
         }
     }
 }
